frequencia/requirements.txt
pdfplumber
pandas
openpyxl
python-dateutil
